pyproject.toml
[project]
name = "grocerytracker-api"
version = "0.1.0"
description = "Grocery Tracker API"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi[standard]>=0.113.0",
    "sqlmodel>=0.0.38",
    "psycopg2>=2.9.11",
    "PyJWT>=2.12.1",
    "pwdlib[argon2]>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.*"
]